Improve code modularity
